feat(api): update User types, move enum
