+++ api/src/main/java/api/database/entities/User.java
@@ -2,6 +2,8 @@ package api.database.entities;
 
 import org.bson.types.ObjectId;
 
+import api.database.enums.AuthProvider;
+
 public class User {
   public ObjectId id;
   public String email;
